# Use AttributeDecoder for unmarshal operations