--- a/attribute_types.go
+++ b/attribute_types.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"time"
 
+	"github.com/mdlayher/netlink"
 	"github.com/pkg/errors"
 	"github.com/ti-mo/netfilter"
 )
@@ -21,101 +22,9 @@ const (
 	opUnSynProxy      = "SynProxy unmarshal"
 )
 
-var (
-	ctaCountersOrigReplyCat = fmt.Sprintf("%s/%s", ctaCountersOrig, ctaCountersReply)
-	ctaSeqAdjOrigReplyCat   = fmt.Sprintf("%s/%s", ctaSeqAdjOrig, ctaSeqAdjReply)
-)
-
-// num16 is a generic numeric attribute. It is represented by a uint32
-// and holds its own AttributeType.
-type num16 struct {
-	Type  attributeType
-	Value uint16
-}
-
-// Filled returns true if the Num16's type is non-zero.
-func (i num16) filled() bool {
-	return i.Type != 0 || i.Value != 0
-}
-
-func (i num16) String() string {
-	return fmt.Sprintf("%d", i.Value)
-}
-
-// unmarshal unmarshals a netfilter.Attribute into a Num16.
-func (i *num16) unmarshal(attr netfilter.Attribute) error {
-
-	if len(attr.Data) != 2 {
-		return errIncorrectSize
-	}
-
-	i.Type = attributeType(attr.Type)
-	i.Value = attr.Uint16()
-
-	return nil
-}
-
-// marshal marshals a Num16 into a netfilter.Attribute. If the AttributeType parameter is non-zero,
-// it is used as Attribute's type; otherwise, the Num16's Type field is used.
-func (i num16) marshal(t attributeType) netfilter.Attribute {
-
-	var nfa netfilter.Attribute
-
-	if t == 0 {
-		nfa.Type = uint16(i.Type)
-	} else {
-		nfa.Type = uint16(t)
-	}
-
-	nfa.PutUint16(i.Value)
-
-	return nfa
-}
-
-// num32 is a generic numeric attribute. It is represented by a uint32
-// and holds its own AttributeType.
-type num32 struct {
-	Type  attributeType
-	Value uint32
-}
-
-// Filled returns true if the Num32's type is non-zero.
-func (i num32) filled() bool {
-	return i.Type != 0 || i.Value != 0
-}
-
-func (i num32) String() string {
-	return fmt.Sprintf("%d", i.Value)
-}
-
-// unmarshal unmarshals a netfilter.Attribute into a Num32.
-func (i *num32) unmarshal(attr netfilter.Attribute) error {
-
-	if len(attr.Data) != 4 {
-		return errIncorrectSize
-	}
-
-	i.Type = attributeType(attr.Type)
-	i.Value = attr.Uint32()
-
-	return nil
-}
-
-// marshal marshals a Num32 into a netfilter.Attribute. If the AttributeType parameter is non-zero,
-// it is used as Attribute's type; otherwise, the Num32's Type field is used.
-func (i num32) marshal(t attributeType) netfilter.Attribute {
-
-	var nfa netfilter.Attribute
-
-	if t == 0 {
-		nfa.Type = uint16(i.Type)
-	} else {
-		nfa.Type = uint16(t)
-	}
-
-	nfa.PutUint32(i.Value)
-
-	return nfa
+// nestedFlag returns true if the NLA_F_NESTED flag is set on typ.
+func nestedFlag(typ uint16) bool {
+	return typ&netlink.Nested != 0
 }
 
 // A Helper holds the name and info the helper that creates a related connection.
@@ -130,28 +39,20 @@ func (hlp Helper) filled() bool {
 }
 
 // unmarshal unmarshals a netfilter.Attribute into a Helper.
-func (hlp *Helper) unmarshal(attr netfilter.Attribute) error {
+func (hlp *Helper) unmarshal(ad *netlink.AttributeDecoder) error {
 
-	if attributeType(attr.Type) != ctaHelp {
-		return fmt.Errorf(errAttributeWrongType, attr.Type, ctaHelp)
-	}
-
-	if !attr.Nested {
-		return errors.Wrap(errNotNested, opUnHelper)
-	}
-
-	for _, iattr := range attr.Children {
-		switch helperType(iattr.Type) {
+	for ad.Next() {
+		switch helperType(ad.Type()) {
 		case ctaHelpName:
-			hlp.Name = string(iattr.Data)
+			hlp.Name = ad.String()
 		case ctaHelpInfo:
-			hlp.Info = iattr.Data
+			hlp.Info = ad.Bytes()
 		default:
-			return fmt.Errorf(errAttributeChild, iattr.Type, ctaHelp)
+			return fmt.Errorf(errAttributeChild, ad.Type())
 		}
 	}
 
-	return nil
+	return ad.Err()
 }
 
 // marshal marshals a Helper into a netfilter.Attribute.
@@ -183,52 +84,40 @@ func (pi ProtoInfo) filled() bool {
 
 // unmarshal unmarshals a netfilter.Attribute into a ProtoInfo structure.
 // one of three ProtoInfo types; TCP, DCCP or SCTP.
-func (pi *ProtoInfo) unmarshal(attr netfilter.Attribute) error {
+func (pi *ProtoInfo) unmarshal(ad *netlink.AttributeDecoder) error {
 
 	// Make sure we don't unmarshal into the same ProtoInfo twice.
 	if pi.filled() {
 		return errReusedProtoInfo
 	}
 
-	if attributeType(attr.Type) != ctaProtoInfo {
-		return fmt.Errorf(errAttributeWrongType, attr.Type, ctaProtoInfo)
-	}
-
-	if !attr.Nested {
-		return errors.Wrap(errNotNested, opUnProtoInfo)
-	}
-
-	if len(attr.Children) != 1 {
+	if ad.Len() != 1 {
 		return errors.Wrap(errNeedSingleChild, opUnProtoInfo)
 	}
 
-	// Step into the single nested child
-	iattr := attr.Children[0]
+	// Step into the single nested child, return on error.
+	if !ad.Next() {
+		return ad.Err()
+	}
 
-	switch protoInfoType(iattr.Type) {
+	switch protoInfoType(ad.Type()) {
 	case ctaProtoInfoTCP:
 		var tpi ProtoInfoTCP
-		if err := tpi.unmarshal(iattr); err != nil {
-			return err
-		}
+		ad.Nested(tpi.unmarshal)
 		pi.TCP = &tpi
 	case ctaProtoInfoDCCP:
 		var dpi ProtoInfoDCCP
-		if err := dpi.unmarshal(iattr); err != nil {
-			return err
-		}
+		ad.Nested(dpi.unmarshal)
 		pi.DCCP = &dpi
 	case ctaProtoInfoSCTP:
 		var spi ProtoInfoSCTP
-		if err := spi.unmarshal(iattr); err != nil {
-			return err
-		}
+		ad.Nested(spi.unmarshal)
 		pi.SCTP = &spi
 	default:
-		return fmt.Errorf(errAttributeChild, iattr.Type, ctaProtoInfo)
+		return fmt.Errorf(errAttributeChild, ad.Type())
 	}
 
-	return nil
+	return ad.Err()
 }
 
 // marshal marshals a ProtoInfo into a netfilter.Attribute.
@@ -258,39 +147,31 @@ type ProtoInfoTCP struct {
 }
 
 // unmarshal unmarshals a netfilter.Attribute into a ProtoInfoTCP.
-func (tpi *ProtoInfoTCP) unmarshal(attr netfilter.Attribute) error {
-
-	if protoInfoType(attr.Type) != ctaProtoInfoTCP {
-		return fmt.Errorf(errAttributeWrongType, attr.Type, ctaProtoInfoTCP)
-	}
-
-	if !attr.Nested {
-		return errors.Wrap(errNotNested, opUnProtoInfoTCP)
-	}
+func (tpi *ProtoInfoTCP) unmarshal(ad *netlink.AttributeDecoder) error {
 
 	// A ProtoInfoTCP has at least 3 members, TCP_STATE and TCP_FLAGS_ORIG/REPLY.
-	if len(attr.Children) < 3 {
+	if ad.Len() < 3 {
 		return errors.Wrap(errNeedChildren, opUnProtoInfoTCP)
 	}
 
-	for _, iattr := range attr.Children {
-		switch protoInfoTCPType(iattr.Type) {
+	for ad.Next() {
+		switch protoInfoTCPType(ad.Type()) {
 		case ctaProtoInfoTCPState:
-			tpi.State = iattr.Data[0]
+			tpi.State = ad.Uint8()
 		case ctaProtoInfoTCPWScaleOriginal:
-			tpi.OriginalWindowScale = iattr.Data[0]
+			tpi.OriginalWindowScale = ad.Uint8()
 		case ctaProtoInfoTCPWScaleReply:
-			tpi.ReplyWindowScale = iattr.Data[0]
+			tpi.ReplyWindowScale = ad.Uint8()
 		case ctaProtoInfoTCPFlagsOriginal:
-			tpi.OriginalFlags = iattr.Uint16()
+			tpi.OriginalFlags = ad.Uint16()
 		case ctaProtoInfoTCPFlagsReply:
-			tpi.ReplyFlags = iattr.Uint16()
+			tpi.ReplyFlags = ad.Uint16()
 		default:
-			return fmt.Errorf(errAttributeChild, iattr.Type, ctaProtoInfoTCP)
+			return fmt.Errorf(errAttributeChild, ad.Type())
 		}
 	}
 
-	return nil
+	return ad.Err()
 }
 
 // marshal marshals a ProtoInfoTCP into a netfilter.Attribute.
@@ -319,34 +200,26 @@ type ProtoInfoDCCP struct {
 }
 
 // unmarshal unmarshals a netfilter.Attribute into a ProtoInfoTCP.
-func (dpi *ProtoInfoDCCP) unmarshal(attr netfilter.Attribute) error {
+func (dpi *ProtoInfoDCCP) unmarshal(ad *netlink.AttributeDecoder) error {
 
-	if protoInfoType(attr.Type) != ctaProtoInfoDCCP {
-		return fmt.Errorf(errAttributeWrongType, attr.Type, ctaProtoInfoDCCP)
-	}
-
-	if !attr.Nested {
-		return errors.Wrap(errNotNested, opUnProtoInfoDCCP)
-	}
-
-	if len(attr.Children) == 0 {
+	if ad.Len() == 0 {
 		return errors.Wrap(errNeedChildren, opUnProtoInfoDCCP)
 	}
 
-	for _, iattr := range attr.Children {
-		switch protoInfoDCCPType(iattr.Type) {
+	for ad.Next() {
+		switch protoInfoDCCPType(ad.Type()) {
 		case ctaProtoInfoDCCPState:
-			dpi.State = iattr.Data[0]
+			dpi.State = ad.Uint8()
 		case ctaProtoInfoDCCPRole:
-			dpi.Role = iattr.Data[0]
+			dpi.Role = ad.Uint8()
 		case ctaProtoInfoDCCPHandshakeSeq:
-			dpi.HandshakeSeq = iattr.Uint64()
+			dpi.HandshakeSeq = ad.Uint64()
 		default:
-			return fmt.Errorf(errAttributeChild, iattr.Type, ctaProtoInfoDCCP)
+			return fmt.Errorf(errAttributeChild, ad.Type())
 		}
 	}
 
-	return nil
+	return ad.Err()
 }
 
 // marshal marshals a ProtoInfoDCCP into a netfilter.Attribute.
@@ -368,34 +241,26 @@ type ProtoInfoSCTP struct {
 }
 
 // unmarshal unmarshals a netfilter.Attribute into a ProtoInfoSCTP.
-func (spi *ProtoInfoSCTP) unmarshal(attr netfilter.Attribute) error {
+func (spi *ProtoInfoSCTP) unmarshal(ad *netlink.AttributeDecoder) error {
 
-	if protoInfoType(attr.Type) != ctaProtoInfoSCTP {
-		return fmt.Errorf(errAttributeWrongType, attr.Type, ctaProtoInfoSCTP)
-	}
-
-	if !attr.Nested {
-		return errors.Wrap(errNotNested, opUnProtoInfoSCTP)
-	}
-
-	if len(attr.Children) == 0 {
+	if ad.Len() == 0 {
 		return errors.Wrap(errNeedChildren, opUnProtoInfoSCTP)
 	}
 
-	for _, iattr := range attr.Children {
-		switch protoInfoSCTPType(iattr.Type) {
+	for ad.Next() {
+		switch protoInfoSCTPType(ad.Type()) {
 		case ctaProtoInfoSCTPState:
-			spi.State = iattr.Data[0]
+			spi.State = ad.Uint8()
 		case ctaProtoInfoSCTPVTagOriginal:
-			spi.VTagOriginal = iattr.Uint32()
+			spi.VTagOriginal = ad.Uint32()
 		case ctaProtoInfoSCTPVtagReply:
-			spi.VTagReply = iattr.Uint32()
+			spi.VTagReply = ad.Uint32()
 		default:
-			return fmt.Errorf(errAttributeChild, iattr.Type, ctaProtoInfoSCTP)
+			return fmt.Errorf(errAttributeChild, ad.Type())
 		}
 	}
 
-	return nil
+	return ad.Err()
 }
 
 // marshal marshals a ProtoInfoSCTP into a netfilter.Attribute.
@@ -438,41 +303,29 @@ func (ctr Counter) filled() bool {
 }
 
 // unmarshal unmarshals a nested counter attribute into a Counter structure.
-func (ctr *Counter) unmarshal(attr netfilter.Attribute) error {
-
-	if attributeType(attr.Type) != ctaCountersOrig &&
-		attributeType(attr.Type) != ctaCountersReply {
-		return fmt.Errorf(errAttributeWrongType, attr.Type, ctaCountersOrigReplyCat)
-	}
-
-	if !attr.Nested {
-		return errors.Wrap(errNotNested, opUnCounter)
-	}
+func (ctr *Counter) unmarshal(ad *netlink.AttributeDecoder) error {
 
 	// A Counter consists of packet and byte attributes but may have
 	// help attributes as well if nf_conntrack_helper enabled
-	if len(attr.Children) < 2 {
+	if ad.Len() < 2 {
 		return errors.Wrap(errNeedChildren, opUnCounter)
 	}
 
-	// Set Direction to true if it's a reply counter
-	ctr.Direction = attributeType(attr.Type) == ctaCountersReply
-
-	for _, iattr := range attr.Children {
-		switch counterType(iattr.Type) {
+	for ad.Next() {
+		switch counterType(ad.Type()) {
 		case ctaCountersPackets:
-			ctr.Packets = iattr.Uint64()
+			ctr.Packets = ad.Uint64()
 		case ctaCountersBytes:
-			ctr.Bytes = iattr.Uint64()
+			ctr.Bytes = ad.Uint64()
 		case ctaCountersPad:
 			// Ignore padding attributes that show up if nf_conntrack_helper is enabled.
 			continue
 		default:
-			return fmt.Errorf(errAttributeChild, iattr.Type, ctaCountersOrigReplyCat)
+			return fmt.Errorf(errAttributeChild, ad.Type())
 		}
 	}
 
-	return nil
+	return ad.Err()
 }
 
 // A Timestamp represents the start and end time of a flow.
@@ -484,33 +337,25 @@ type Timestamp struct {
 }
 
 // unmarshal unmarshals a nested timestamp attribute into a conntrack.Timestamp structure.
-func (ts *Timestamp) unmarshal(attr netfilter.Attribute) error {
-
-	if attributeType(attr.Type) != ctaTimestamp {
-		return fmt.Errorf(errAttributeWrongType, attr.Type, ctaTimestamp)
-	}
-
-	if !attr.Nested {
-		return errors.Wrap(errNotNested, opUnTimestamp)
-	}
+func (ts *Timestamp) unmarshal(ad *netlink.AttributeDecoder) error {
 
 	// A Timestamp will always have at least a start time
-	if len(attr.Children) == 0 {
+	if ad.Len() == 0 {
 		return errors.Wrap(errNeedSingleChild, opUnTimestamp)
 	}
 
-	for _, iattr := range attr.Children {
-		switch timestampType(iattr.Type) {
+	for ad.Next() {
+		switch timestampType(ad.Type()) {
 		case ctaTimestampStart:
-			ts.Start = time.Unix(0, iattr.Int64())
+			ts.Start = time.Unix(0, int64(ad.Uint64()))
 		case ctaTimestampStop:
-			ts.Stop = time.Unix(0, iattr.Int64())
+			ts.Stop = time.Unix(0, int64(ad.Uint64()))
 		default:
-			return fmt.Errorf(errAttributeChild, iattr.Type, ctaTimestamp)
+			return fmt.Errorf(errAttributeChild, ad.Type())
 		}
 	}
 
-	return nil
+	return ad.Err()
 }
 
 // A Security structure holds the security info belonging to a connection.
@@ -519,31 +364,23 @@ func (ts *Timestamp) unmarshal(attr netfilter.Attribute) error {
 type Security string
 
 // unmarshal unmarshals a nested security attribute into a conntrack.Security structure.
-func (sec *Security) unmarshal(attr netfilter.Attribute) error {
-
-	if attributeType(attr.Type) != ctaSecCtx {
-		return fmt.Errorf(errAttributeWrongType, attr.Type, ctaSecCtx)
-	}
-
-	if !attr.Nested {
-		return errors.Wrap(errNotNested, opUnSecurity)
-	}
+func (sec *Security) unmarshal(ad *netlink.AttributeDecoder) error {
 
 	// A SecurityContext has at least a name
-	if len(attr.Children) == 0 {
+	if ad.Len() == 0 {
 		return errors.Wrap(errNeedChildren, opUnSecurity)
 	}
 
-	for _, iattr := range attr.Children {
-		switch securityType(iattr.Type) {
+	for ad.Next() {
+		switch securityType(ad.Type()) {
 		case ctaSecCtxName:
-			*sec = Security(iattr.Data)
+			*sec = Security(ad.Bytes())
 		default:
-			return fmt.Errorf(errAttributeChild, iattr.Type, ctaSecCtx)
+			return fmt.Errorf(errAttributeChild, ad.Type())
 		}
 	}
 
-	return nil
+	return ad.Err()
 }
 
 // SequenceAdjust represents a TCP sequence number adjustment event.
@@ -575,39 +412,27 @@ func (seq SequenceAdjust) filled() bool {
 
 // unmarshal unmarshals a nested sequence adjustment attribute into a
 // conntrack.SequenceAdjust structure.
-func (seq *SequenceAdjust) unmarshal(attr netfilter.Attribute) error {
-
-	if attributeType(attr.Type) != ctaSeqAdjOrig &&
-		attributeType(attr.Type) != ctaSeqAdjReply {
-		return fmt.Errorf(errAttributeWrongType, attr.Type, ctaSeqAdjOrigReplyCat)
-	}
-
-	if !attr.Nested {
-		return errors.Wrap(errNotNested, opUnSeqAdj)
-	}
+func (seq *SequenceAdjust) unmarshal(ad *netlink.AttributeDecoder) error {
 
 	// A SequenceAdjust message should come with at least 1 child.
-	if len(attr.Children) == 0 {
+	if ad.Len() == 0 {
 		return errors.Wrap(errNeedSingleChild, opUnSeqAdj)
 	}
 
-	// Set Direction to true if it's a reply adjustment
-	seq.Direction = attributeType(attr.Type) == ctaSeqAdjReply
-
-	for _, iattr := range attr.Children {
-		switch seqAdjType(iattr.Type) {
+	for ad.Next() {
+		switch seqAdjType(ad.Type()) {
 		case ctaSeqAdjCorrectionPos:
-			seq.Position = iattr.Uint32()
+			seq.Position = ad.Uint32()
 		case ctaSeqAdjOffsetBefore:
-			seq.OffsetBefore = iattr.Uint32()
+			seq.OffsetBefore = ad.Uint32()
 		case ctaSeqAdjOffsetAfter:
-			seq.OffsetAfter = iattr.Uint32()
+			seq.OffsetAfter = ad.Uint32()
 		default:
-			return fmt.Errorf(errAttributeChild, iattr.Type, ctaSeqAdjOrigReplyCat)
+			return fmt.Errorf(errAttributeChild, ad.Type())
 		}
 	}
 
-	return nil
+	return ad.Err()
 }
 
 // marshal marshals a SequenceAdjust into a netfilter.Attribute.
@@ -642,34 +467,26 @@ func (sp SynProxy) filled() bool {
 }
 
 // unmarshal unmarshals a SYN proxy attribute into a SynProxy structure.
-func (sp *SynProxy) unmarshal(attr netfilter.Attribute) error {
+func (sp *SynProxy) unmarshal(ad *netlink.AttributeDecoder) error {
 
-	if attributeType(attr.Type) != ctaSynProxy {
-		return fmt.Errorf(errAttributeWrongType, attr.Type, ctaSynProxy)
-	}
-
-	if !attr.Nested {
-		return errors.Wrap(errNotNested, opUnSynProxy)
-	}
-
-	if len(attr.Children) == 0 {
+	if ad.Len() == 0 {
 		return errors.Wrap(errNeedSingleChild, opUnSynProxy)
 	}
 
-	for _, iattr := range attr.Children {
-		switch synProxyType(iattr.Type) {
+	for ad.Next() {
+		switch synProxyType(ad.Type()) {
 		case ctaSynProxyISN:
-			sp.ISN = iattr.Uint32()
+			sp.ISN = ad.Uint32()
 		case ctaSynProxyITS:
-			sp.ITS = iattr.Uint32()
+			sp.ITS = ad.Uint32()
 		case ctaSynProxyTSOff:
-			sp.TSOff = iattr.Uint32()
+			sp.TSOff = ad.Uint32()
 		default:
-			return fmt.Errorf(errAttributeChild, iattr.Type, ctaSynProxy)
+			return fmt.Errorf(errAttributeChild, ad.Type())
 		}
 	}
 
-	return nil
+	return ad.Err()
 }
 
 // marshal marshals a SynProxy into a netfilter.Attribute.