--- a/go.mod
+++ b/go.mod
@@ -2,12 +2,14 @@ module github.com/ti-mo/conntrack
 
 go 1.12
 
+replace github.com/ti-mo/netfilter => /home/timo/documents/go/netfilter
+
 require (
-	github.com/google/go-cmp v0.2.0
-	github.com/mdlayher/netlink v0.0.0-20190313131330-258ea9dff42c
+	github.com/google/go-cmp v0.3.1
+	github.com/mdlayher/netlink v1.0.1-0.20191210152442-a1644773bc99
 	github.com/pkg/errors v0.8.1
 	github.com/stretchr/testify v1.4.0
 	github.com/ti-mo/netfilter v0.2.0
 	github.com/vishvananda/netns v0.0.0-20180720170159-13995c7128cc
-	golang.org/x/sys v0.0.0-20190322080309-f49334f85ddc
+	golang.org/x/sys v0.0.0-20191210023423-ac6580df4449
 )