Flows now potentially unmarshal an order of magnitude faster by using `netlink.AttributeDecoder`. This removes all intermediate allocations from the hot path.

Before:
```
λ  conntrack master ✓ go test -bench=.
goos: linux
goarch: amd64
pkg: github.com/ti-mo/conntrack
BenchmarkExpectUnmarshal-16             	 1316797	      1001 ns/op	     144 B/op	      10 allocs/op
BenchmarkFlowUnmarshal-16               	  998860	      1231 ns/op	     120 B/op	       9 allocs/op
PASS
ok  	github.com/ti-mo/conntrack	4.513s
```

After:
```
λ  conntrack netlink-attributedecoder ✓ go test -bench=.
goos: linux
goarch: amd64
pkg: github.com/ti-mo/conntrack
BenchmarkExpectUnmarshal-16             	10596650	       113 ns/op	     320 B/op	       1 allocs/op
BenchmarkFlowUnmarshal-16               	 7802109	       151 ns/op	     512 B/op	       1 allocs/op
BenchmarkStatusUnmarshalAttribute-16    	11117895	        99.6 ns/op
PASS
ok  	github.com/ti-mo/conntrack	3.887s
```

@smarterclayton While this is not exactly what you're asking for in https://github.com/ti-mo/conntrack/issues/13, it's a step closer to where the library needs to be with regards to performance. The constants are not public yet, nor is there a method that returns an `AttributeDecoder` directly. However, if you need something `now()`, check out the usage of `netfilter.DecodeNetlink()`.

cc @mdlayher